moti/requirements.txt
Django
django-cors-headers
djangorestframework
django_extensions

